+++ src/interpret.c
@@ -545,7 +545,8 @@ void dump_backlog(void)
       {
 	fprintf(stderr,"(%ld)", (long)backlog[e].arg);
       }
-      fprintf(stderr," %d, %d\n", backlog[e].stack, backlog[e].mark_stack);
+      fprintf(stderr," %ld, %ld\n",
+	      (long)backlog[e].stack, (long)backlog[e].mark_stack);
     }
   }while(e!=backlogp);
 }
